update tests and router
