+++ sensorsafrica/api/v1/views.py
@@ -1,4 +1,5 @@
 import django_filters
+from django.db import models
 from feinstaub.sensors.models import SensorData
 from feinstaub.sensors.views import SensorDataView as FeinstaubSensorDataView
 
@@ -7,6 +8,14 @@ class SensorFilter(django_filters.FilterSet):
     class Meta:
         model = SensorData
         fields = {"sensor": ["exact"], "timestamp": ["gte"]}
+        filter_overrides = {
+             models.DateTimeField: {
+                 'filter_class': django_filters.IsoDateTimeFilter,
+                 'extra': lambda f: {
+                     'lookup_expr': 'gte',
+                 },
+             },
+         }
 
 
 class SensorDataView(FeinstaubSensorDataView):
